Optimizations
All MSP computations now use the new sampling procedure, which is
faster.
And all computations use on the fly sampling, meaning that a sample is
created, the MSP values are computed and only then a next sample is
created. This allows for much bigger and much more samples, without
memory problems.