Speedup
The weighted versions of MSP were substantially slower than the
unweighted versions, because computing the relative frequency of a
lemma pretty expensive (it computed n(., ., age) over and over again).
I replaced f by fSpeed in which n(., ., age) is provided as an input
parameter. Because f(*, lemma, age) is called as many times as there
are lemmas at a certain age, this saves a considerable amount of work.